Commit current version of kernel
All work done so far is just to manage a basic boot and check the values
currently stored within the info given by multiboot. The kernel is
higher-half, so, unfortunately, nothing has been done to make the
information <1MiB that it ay point to accessible. This is the next step
to be completed.

The console routines as given operate on VGA. As this is hopelessly
out of date, I would prefer to switch over to detecting GOP and VESA
first, and that would be the next step for the kernel itself.